added unit tests for 'Use_ANSI' feature. Some minor changes to satisfy lint
