+++ pkg/integrations/v4/logs/cfg_others.go
@@ -2,13 +2,12 @@
 // +build !windows
 
 /*
- * Copyright 2021 New Relic Corporation. All rights reserved.
+ * Copyright 2023 New Relic Corporation. All rights reserved.
  * SPDX-License-Identifier: Apache-2.0
  */
 
 package logs
 
 func addOSDependantConfig(fbOSConf FBOSConfig) FBOSConfig {
-	fbOSConf.ForceUseANSI = true
 	return fbOSConf
 }
